fix(gatsby-theme-docz): replace iframe in playground with div #984 #1035
